requirements.txt
numba==0.64.0
numpy==2.4.3
pandas==3.0.1
plotly==6.5.0
protobuf==7.34.0
PyYAML==6.0.3
streamlit==1.55.0
tqdm==4.67.3
baostock>=0.8.13
zhipuai>=2.0.0
